dedupe files
